requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
regex==2023.12.25
aiofiles==23.2.1
asyncio==3.4.3
